@@ -1,67 +1,63 @@
 {
+  "name": "remix-starter-template",
   "private": true,
   "sideEffects": false,
+  "type": "module",
   "author": {
     "name": "Franklin Javier",
     "email": "franklinjalves@gmail.com"
   },
   "scripts": {
-    "build:css": "npm run generate:css -- --minify",
-    "build:remix": "remix build",
-    "build": "run-s \"build:*\"",
-    "dev:css": "npm run generate:css -- --watch",
-    "dev:remix": "remix dev",
-    "dev": "run-p \"dev:*\"",
-    "format": "prettier --write .",
-    "generate:css": "npx tailwindcss -o ./app/tailwind.css",
+    "build": "remix build",
+    "dev": "remix dev --manual",
+    "start": "remix-serve ./build/index.js",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "clean": "rm -rf node_modules pnpm-lock.yaml",
+    "format": "prettier --write \"**/*.{ts,tsx,md}\"",
+    "typecheck": "tsc",
     "generate": "plop",
-    "lint:ci": "eslint --ignore-path .gitignore .",
-    "lint:fix": "eslint --cache --ignore-path .gitignore --fix .",
-    "lint": "eslint --cache --ignore-path .gitignore .",
     "prepare": "husky install",
-    "start": "remix-serve build",
     "test:dev": "jest --watchAll",
     "test": "jest --forceExit --no-cache",
     "type-check": "tsc",
     "updatedeps": "yarn --latest upgrade-interactive"
   },
   "dependencies": {
-    "@remix-run/node": "^1.8.2",
-    "@remix-run/react": "^1.8.2",
-    "@remix-run/serve": "^1.8.2",
-    "@remix-run/vercel": "^1.8.2",
-    "clsx": "1.2.1",
-    "isbot": "^3.6.5",
+    "@remix-run/node": "^2.1.0",
+    "@remix-run/react": "^2.1.0",
+    "@remix-run/serve": "^2.1.0",
+    "clsx": "2.0.0",
+    "isbot": "^3.7.0",
+    "lucide-react": "^0.290.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@remix-run/dev": "^1.8.2",
-    "@remix-run/eslint-config": "^1.8.2",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^14.4.3",
-    "@types/jest": "^29.2.4",
-    "@types/react": "^18.0.26",
-    "@types/react-dom": "^18.0.9",
-    "eslint": "^8.29.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-import": "^2.26.0",
+    "@remix-run/dev": "^2.1.0",
+    "@remix-run/eslint-config": "^2.1.0",
+    "@testing-library/jest-dom": "^6.1.4",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.5.1",
+    "@types/jest": "^29.5.6",
+    "@types/react": "^18.2.33",
+    "@types/react-dom": "^18.2.14",
+    "eslint": "^8.52.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-import": "^2.29.0",
     "git-cz": "^4.9.0",
-    "husky": "^8.0.2",
-    "jest": "29.3.1",
-    "jest-environment-jsdom": "^29.3.1",
-    "lint-staged": "^13.1.0",
-    "npm-run-all": "^4.1.5",
-    "plop": "^3.1.1",
-    "prettier": "^2.8.1",
-    "prettier-plugin-tailwindcss": "^0.2.1",
-    "tailwindcss": "^3.2.4",
-    "ts-jest": "29.0.3",
-    "typescript": "^4.9.4"
+    "husky": "^8.0.3",
+    "jest": "29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "lint-staged": "^15.0.2",
+    "plop": "^4.0.0",
+    "prettier": "^3.0.3",
+    "prettier-plugin-tailwindcss": "^0.5.6",
+    "tailwindcss": "^3.3.5",
+    "ts-jest": "29.1.1",
+    "typescript": "^5.2.2"
   },
   "engines": {
-    "node": ">=16"
+    "node": ">=18.0.0"
   },
   "config": {
     "commitizen": {
