@@ -1,8 +1,5 @@
-/**
- * @type {import('@remix-run/dev').AppConfig}
- */
-module.exports = {
-  serverBuildTarget: 'vercel',
-  server: process.env.NODE_ENV === 'development' ? undefined : './server.js',
-  ignoredRouteFiles: ['.*'],
+/** @type {import('@remix-run/dev').AppConfig} */
+export default {
+  ignoredRouteFiles: ['**/.*'],
+  tailwind: true,
 }
