@@ -1,25 +1,41 @@
 import { Links, LiveReload, Meta, Outlet, Scripts, ScrollRestoration } from '@remix-run/react'
 
-import styles from './tailwind.css'
+import styles from './globals.css'
 
-import type { LinksFunction, MetaFunction } from '@remix-run/node'
+import type { LinksFunction, LoaderFunctionArgs } from '@remix-run/node'
 
-export const meta: MetaFunction = () => ({
-  charset: 'utf-8',
-  title: 'New Remix App',
-  viewport: 'width=device-width,initial-scale=1',
-})
+export const links: LinksFunction = () => [
+  { rel: 'stylesheet', href: styles },
+  {
+    rel: 'icon',
+    href: '/logo.svg',
+    type: 'image/svg',
+  },
+]
 
-export const links: LinksFunction = () => [{ rel: 'stylesheet', href: styles }]
+export async function loader({ request }: LoaderFunctionArgs) {
+  return true
+}
 
 export default function App() {
   return (
     <html lang="en">
       <head>
+        <meta charSet="utf-8" />
+        <meta content="width=device-width, initial-scale=1" name="viewport" />
         <Meta />
         <Links />
       </head>
       <body>
+        {/**
+         * This removes anything added to html from extensions, causing hydration issue
+          https://github.com/remix-run/remix/issues/4822
+        */}
+        <script
+          dangerouslySetInnerHTML={{
+            __html: `document.querySelectorAll("html > script").forEach((s) => s.parentNode?.removeChild(s));`,
+          }}
+        />
         <Outlet />
         <ScrollRestoration />
         <Scripts />
